cloud_functions/ingest-ditat/requirements.txt
requests>=2.32.3
functions-framework==3.8.2
flask==3.0.3
google-cloud-error-reporting==1.11.1
MarkupSafe==2.1.3
pandas==2.3.0
pandas-gbq==0.29.1
python-dotenv==1.1.1
